+++ src/main/java/com/unascribed/yttr/content/enchant/DisjunctionEnchantment.java
@@ -1,5 +1,7 @@
 package com.unascribed.yttr.content.enchant;
 
+import com.unascribed.yttr.YConfig;
+
 import net.minecraft.enchantment.DamageEnchantment;
 import net.minecraft.entity.EntityGroup;
 import net.minecraft.entity.EquipmentSlot;
@@ -10,6 +12,18 @@ public class DisjunctionEnchantment extends DamageEnchantment {
 	
 	public DisjunctionEnchantment() {
 		super(Rarity.UNCOMMON, 1, EquipmentSlot.MAINHAND);
+	}
+
+	@Override
+	public int getMinPower(int level) {
+		if (!YConfig.Enchantments.disjunction) return 30000;
+		return super.getMinPower(level);
+	}
+
+	@Override
+	public int getMaxPower(int level) {
+		if (!YConfig.Enchantments.disjunction) return -30000;
+		return super.getMaxPower(level);
 	}
 	
 	@Override