+++ src/main/java/com/unascribed/yttr/content/enchant/StabilizationEnchantment.java
@@ -1,5 +1,6 @@
 package com.unascribed.yttr.content.enchant;
 
+import com.unascribed.yttr.YConfig;
 import com.unascribed.yttr.init.YEnchantments;
 
 import net.minecraft.enchantment.Enchantment;
@@ -11,11 +12,13 @@ public class StabilizationEnchantment extends CoilEnchantment {
 
 	@Override
 	public int getMinPower(int level) {
+		if (!YConfig.Enchantments.stabilization) return 30000;
 		return 15;
 	}
 
 	@Override
 	public int getMaxPower(int level) {
+		if (!YConfig.Enchantments.stabilization) return -30000;
 		return super.getMinPower(level) + 50;
 	}